Program that will give the user the area code of telephone number
This program is in the basic level as i am going to add more functions to it, it once provided with a telephone number in any format whether you have ( - . space _ between numbers it will convert it to a 10 digits number then it will return the location of the number according to the area code. valid for United states numbers only!